Rename subgraph.py > subgraph_alg.py
