+++ networkx/algorithms/centrality/__init__.py
@@ -1,7 +1,7 @@
 from .betweenness import *
 from .betweenness_subset import *
 from .closeness import *
-from .subgraph import *
+from .subgraph_alg import *
 from .current_flow_closeness import *
 from .current_flow_betweenness import *
 from .current_flow_betweenness_subset import *
